% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCalc.R
\name{branchCalc}
\alias{branchCalc}
\title{Branch Calculation}
\usage{
branchCalc(
  data,
  eval = "ung_2",
  dbh,
  height = NULL,
  species,
  appearance = NULL,
  crown_cond,
  output = "biomass",
  decay = TRUE,
  plot_radius = 11.28
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{eval}{Which allomentic equation should be used? Default is "ung_2". Options are:
"ung_1", "ung_2", "lambert_1", "lambert_2".}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Optional. Required for lambert_2 or ung_2.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{appearance}{Optional. Required when decay = TRUE.}

\item{crown_cond}{Column within data where crown condition is specified.}

\item{output}{One of "biomass" (kg, default), "biomass Mg/ha", "carbon" (kg), or "carbon Mg/ha".}

\item{decay}{Logical with default = TRUE. Should the decay class reduction factor be applied?}

\item{plot_radius}{Plot radius in metres. Required when output is "biomass Mg/ha" or "carbon Mg/ha".
Must be one of 3.99, 5.64, 7.98, or 11.28. Default 11.28 m.}
}
\value{
A vector.
}
\description{
Calculates branch biomass or carbon per tree given either DBH or DBH and Height.
Calculations are done base on the allometric equations provided in Ung et al., 2008 or Lambert et al., 2005.
A species specific decay class reduction factor can be applied (Harmon et al., 2011) if desired.
}
\examples{
branchCalc(data = trees_data, eval = "ung_2", species = "LGTREE_NFI",
dbh = "DBH", height = "HEIGHT", appearance = "APPEARANCE",
crown_cond = "CROWN_COND", output = "biomass", decay = TRUE)
}
