% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foliageCalculator.R
\name{foliageCalculator}
\alias{foliageCalculator}
\title{Internal Function: Foliage Calculator helper function}
\usage{
foliageCalculator(
  data,
  method,
  output,
  dbh,
  height = NULL,
  species,
  crown_cond,
  func,
  appearance = NULL,
  decay = TRUE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{method}{Dataframe within the package to retrieve beta values from.}

\item{output}{Either "biomass" or "carbon".}

\item{dbh}{Column within data where DBH is specified.}

\item{height}{Column within data where Height is specified.}

\item{species}{Column within data where species is specified.}

\item{crown_cond}{Column within data where crown condition is specified.}

\item{func}{Calls function ungEqn.}

\item{appearance}{Column within data where tree appearance is specified.}

\item{decay}{Logical. Should the decay class reduction factor be applied?}
}
\value{
A vector
}
\description{
Internal Function: Foliage Calculator helper function
}
\examples{
NA
}
\keyword{internal}
