% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JENKINS.R
\docType{data}
\name{JENKINS}
\alias{JENKINS}
\title{Data: Jenkins et al. (2003) total above-ground biomass equation coefficients}
\format{
A data frame with 11 rows and 3 columns:
\describe{
\item{SPECIES_GROUP}{Jenkins species group name}
\item{BETA_0}{Intercept coefficient}
\item{BETA_1}{Slope coefficient}
}
}
\source{
Values from Jenkins et al. (2003) National-scale biomass estimators for
United States tree species. Forest Science, 49(1): 12-35. "Mixed softwood" is
an in-house addition: arithmetic mean of beta coefficients across Cedar/larch,
Douglas-fir, True fir/hemlock, Pine, and Spruce.
}
\usage{
JENKINS
}
\description{
Data: Jenkins et al. (2003) total above-ground biomass equation coefficients
}
\details{
Used in the equation \code{bm = exp(BETA_0 + BETA_1 * log(DBH))}, where DBH is
in cm and bm is total above-ground biomass in kg.
}
\keyword{datasets}
