% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkinsRatioCalculator.R
\name{jenkinsRatioCalculator}
\alias{jenkinsRatioCalculator}
\title{Internal Function: Jenkins Ratio Calculator}
\usage{
jenkinsRatioCalculator(data, dbh, species)
}
\arguments{
\item{data}{User specified dataframe.}

\item{dbh}{Column within data where DBH is specified (cm).}

\item{species}{Column within data where species is specified. NFI codes.}
}
\value{
A named list of four numeric vectors: foliage, coarse_roots, stem_bark, stem_wood.
}
\description{
Internal Function: Jenkins Ratio Calculator
}
\examples{
NA
}
\keyword{internal}
