% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAMBERT_1.R
\docType{data}
\name{LAMBERT_1}
\alias{LAMBERT_1}
\title{Coefficients for allometric equations for DBH only}
\format{
A data frame with 288 rows and 4 columns:
\describe{
\item{SPECIES}{Species codes. Follows NFI naming conventions}
\item{COMPONENT}{Part of the tree}
\item{BETA_NUM}{Corrresponds to beta coefficents}
\item{VALUES}{Values}
}
}
\source{
Values from Lambert et al., 2005 Canadian national tree aboveground biomass equations
}
\usage{
LAMBERT_1
}
\description{
Coefficients for allometric equations for DBH only
}
\keyword{datasets}
