% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAMBERT_2.R
\docType{data}
\name{LAMBERT_2}
\alias{LAMBERT_2}
\title{Coefficients for allometric equations for DBH and HEIGHT}
\format{
A data frame with 432 rows and 4 columns:
\describe{
\item{SPECIES}{Species codes. Follows NFI naming conventions}
\item{COMPONENT}{Part of the tree}
\item{BETA_NUM}{Corrresponds to beta coefficents}
\item{VALUES}{Values}
}
}
\source{
Values from Lambert et al., 2005 Canadian national tree aboveground biomass equations
}
\usage{
LAMBERT_2
}
\description{
Coefficients for allometric equations for DBH and HEIGHT
}
\keyword{datasets}
