% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootCalculator.R
\name{rootCalculator}
\alias{rootCalculator}
\title{Internal Function: Root Biomass Calculator}
\usage{
rootCalculator(data, biomass, species, root_type, output)
}
\arguments{
\item{data}{User specified dataframe.}

\item{biomass}{Column within data where pre-computed above-ground biomass (Mg/ha) is specified.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{root_type}{One of "total", "fine", or "coarse".}

\item{output}{One of "biomass Mg/ha" or "carbon Mg/ha".}
}
\value{
A numeric vector.
}
\description{
Internal Function: Root Biomass Calculator
}
\examples{
NA
}
\keyword{internal}
