% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeCalc.R
\name{treeCalc}
\alias{treeCalc}
\title{Tree Biomass Calculation}
\usage{
treeCalc(
  data,
  eval = "ung_eqn_2",
  dbh,
  height = NULL,
  species,
  appearance = NULL,
  crown_cond,
  rem_bark = NULL,
  output = "biomass",
  decay = TRUE,
  include_root = FALSE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{eval}{Which allometric equation should be used? Default is "ung_eqn_2". Options are:
"ung_eqn_1", "ung_eqn_2", "lambert_eqn_1", "lambert_eqn_2".}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Optional. Required for lambert_eqn_2 or ung_eqn_2.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{appearance}{Optional. Required when decay = TRUE.}

\item{crown_cond}{Column within data where crown condition is specified.}

\item{rem_bark}{Optional. Column within data where remaining bark percentage (0-100) is specified.}

\item{output}{Either "biomass" (default, kg) or "carbon" (Mg/ha).}

\item{decay}{Logical with default = TRUE. Should the decay class reduction factor be applied?}

\item{include_root}{Logical with default = FALSE. Should root biomass be included?
rootCalc() is not yet implemented; setting this to TRUE will raise an error.}
}
\value{
A vector.
}
\description{
Calculates total above-ground tree biomass or carbon by summing
wood, branch, foliage, and bark components. Calculations are based on the
allometric equations provided in Ung et al., 2008 or Lambert et al., 2005.
A species-specific decay class reduction factor (Harmon et al., 2011) can be applied.
}
\examples{
treeCalc(data = trees_data, eval = "ung_eqn_2", species = "LGTREE_NFI",
dbh = "DBH", height = "HEIGHT", appearance = "APPEARANCE",
crown_cond = "CROWN_COND", output = "biomass", decay = TRUE)
}
