% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNG_2.R
\docType{data}
\name{UNG_2}
\alias{UNG_2}
\title{Data: Coefficients for allometric equations for DBH and HEIGHT}
\format{
A data frame with 204 rows and 4 columns:
\describe{
\item{SPECIES}{Species codes. Follows NFI naming conventions}
\item{COMPONENT}{Part of the tree}
\item{BETA_NUM}{Corrresponds to beta coefficents}
\item{VALUES}{Values}
}
}
\source{
Values from Ung et al., 2008 Canadian national biomass equations: new parameter estimates that include British Columbia data
}
\usage{
UNG_2
}
\description{
Data: Coefficients for allometric equations for DBH and HEIGHT
}
\keyword{datasets}
