% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungEqn2.R
\name{ungEqn2}
\alias{ungEqn2}
\title{Internal Function: ungEqn2}
\usage{
ungEqn2(data, dbh, height, beta_list, i)
}
\arguments{
\item{data}{User specified dataframe.}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Column within data where Height is specified.}

\item{beta_list}{specified from parent function}

\item{i}{for iterating}
}
\value{
value (numeric)
}
\description{
Internal Function: ungEqn2
}
\examples{
NA
}
