% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkCalc.R
\name{barkCalc}
\alias{barkCalc}
\title{Bark Calculation}
\usage{
barkCalc(
  data,
  eval = "ung_eqn_2",
  dbh,
  height = NULL,
  species,
  appearance = NULL,
  rem_bark = NULL,
  output = "biomass",
  decay = TRUE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{eval}{Which allomentic equation should be used? Default is "ung_eqn_2". Options are:
"ung_eqn_1", "ung_eqn_2", "lambert_eqn_1", "lambert_eqn_2".}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Optional. Required for lambert_eqn_2 or ung_eqn_2.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{appearance}{Optional. Required when decay = TRUE.}

\item{rem_bark}{Optional. Column within data where remaining bark percentage (0-100) is specified.}

\item{output}{Either "biomass" (default, kg) or "carbon" (Mg/ha).}

\item{decay}{Logical with default = TRUE. Should the decay class reduction factor be applied?}
}
\value{
A vector.
}
\description{
Calculates bark biomass or carbon per tree given either DBH or DBH and Height.
Calculations are done base on the allometric equations provided in Ung et al., 2008 or Lambert et al., 2005.
A species specific decay class reduction factor can be applied (Harmon et al., 2011) if desired.
}
\examples{
barkCalc(data = trees_data, eval = "ung_eqn_2", species = "LGTREE_NFI",
dbh = "DBH", height = "HEIGHT", appearance = "APPEARANCE",
rem_bark = "REM_BARK", output = "biomass", decay = TRUE)
}
