% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkCalculator.R
\name{barkCalculator}
\alias{barkCalculator}
\title{Internal Function: Bark Calculator helper function}
\usage{
barkCalculator(
  data,
  method,
  output,
  dbh,
  height = NULL,
  species,
  rem_bark = NULL,
  func,
  appearance = NULL,
  decay = TRUE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{method}{Dataframe within the package to retrieve beta values from.}

\item{output}{Either "biomass" or "carbon".}

\item{dbh}{Column within data where DBH is specified.}

\item{height}{Column within data where Height is specified.}

\item{species}{Column within data where species is specified.}

\item{rem_bark}{Column within data where remaining bark percentage (0-100) is specified.}

\item{func}{Calls function ungEqn.}

\item{appearance}{Column within data where tree appearance is specified.}

\item{decay}{Logical. Should the decay class reduction factor be applied?}
}
\value{
A vector
}
\description{
Internal Function: Bark Calculator helper function
}
\examples{
NA
}
