% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaVal.R
\name{betaVal}
\alias{betaVal}
\title{Internal Function: Beta Values}
\usage{
betaVal(method, species = species_spec, component = component_spec)
}
\arguments{
\item{method}{Dataframe within the package to retrive beta values from.
Options are "LAMBERT_1", "LAMBERT_2", "UNG_1", "UNG_2".}

\item{species}{gets specifed from parent function. Will be an NFI code}

\item{component}{one of "WOOD", "BARK", "BRANCHES" or "FOLIAGE".}
}
\value{
A list of 2 or 3 beta values
}
\description{
Internal Function: Beta Values
}
\examples{
NA
}
