% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCalc.R
\name{branchCalc}
\alias{branchCalc}
\title{Branch Calculation}
\usage{
branchCalc(
  data,
  eval = "ung_eqn_2",
  dbh,
  height,
  appearance,
  species,
  crown_cond,
  output = "biomass",
  decay = TRUE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{eval}{Which allomentic equation should be used? Default is set to Ung Equation 2 where biomass is calculated
with both DBH and Height. Options are: "ung_eqn_1" for DBH based calculations, "ung_eqn_2" for DBH and Height based calculations,
"lambert_eqn_1" for DBH based calculations, and "lambert_eqn_2" for DBH and Height based calculations. Use Ung equations for
within British Columbia and use Lambert equations for species found in Eastern Canada.}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Column within data where Height is specified.}

\item{appearance}{Column within data where tree appearance is specified.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{crown_cond}{Column within data where crown condition is specified.}

\item{output}{Either "biomass" (default) or "carbon". Biomass gives values in Kg. Carbon gives values in Mg/ha.}

\item{decay}{Logical with default = TRUE. Should the decay class reduction factor be applied?}
}
\value{
A vector.
}
\description{
Calculates branch biomass or carbon per tree given either DBH or DBH and Height.
Calculations are done base on the allometric equations provided in Ung et al., 2008 or Lambert et al., 2005.
A species specific decay class reduction factor can be applied (Harmon et al., 2011) if desired.
}
\examples{
branchCalc(data = trees_data,
eval = "ung_eqn_2",
species = "LGTREE_NFI",
dbh = "DBH",
height = "HEIGHT",
appearance = "APPEARANCE",
crown_cond = "CROWN_COND",
output = "biomass",
decay = TRUE)


}
