% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchCalculator.R
\name{branchCalculator}
\alias{branchCalculator}
\title{Internal Function: Branch Calculator helper function}
\usage{
branchCalculator(
  data,
  method,
  output,
  dbh,
  height,
  species,
  crown_cond,
  appearance,
  func,
  decay = TRUE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{method}{Dataframe within the package to retrive beta values from.
Options are "LAMBERT_1", "LAMBERT_2", "UNG_1", "UNG_2".}

\item{output}{Either "biomass" or "carbon". Biomass gives values in Kg. Carbon gives values in Mg/ha.}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Column within data where Height is specified.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{crown_cond}{Column within data where crown condition is specified.}

\item{appearance}{Column within data where tree appearance is specified.}

\item{func}{Calls function ungEqn1 or ungEqn2.}

\item{decay}{Logical with default = TRUE. Should the decay class reduction factor be applied?}
}
\value{
A vector
}
\description{
Internal Function: Branch Calculator helper function
}
\examples{
NA
}
