% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_data.R
\docType{data}
\name{carbon_data}
\alias{carbon_data}
\title{Data: Biomass or carbon modification parameter}
\format{
A data frame with 2 rows and 2 columns:
\describe{
\item{CARBON_METHOD}{Output is "carbon" or "biomass"}
\item{CARBON_MOD}{Value for modification. "carbon"  = 0.5*25/1000}
}
}
\source{
Created in house
}
\usage{
carbon_data
}
\description{
Data: Biomass or carbon modification parameter
}
\keyword{datasets}
