% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonMod.R
\name{carbonMod}
\alias{carbonMod}
\title{Biomass/Carbon output selection}
\usage{
carbonMod(output)
}
\arguments{
\item{output}{"biomass" or "carbon"}
}
\value{
A parameter to modify function output to return either biomass (Kg) or carbon (Mg/ha)
}
\description{
Biomass/Carbon output selection
}
\examples{
carbonMod("biomass")
}
