% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCRF.R
\name{DCRF}
\alias{DCRF}
\title{Decay Class Reduction Factor (DCRF)}
\usage{
DCRF(data, appearance, species, i)
}
\arguments{
\item{data}{User dataframe.}

\item{appearance}{Column from user dataframe where tree appearance is specified.
Appearance codes correspond to Field Manual for Describing Terrestrial Ecosystems 2nd edition.}

\item{species}{Column from user dataframe with species code. Species code must be in NFI format.}

\item{i}{Used for iterating.}
}
\value{
A decay modifer (numeric) corresponding to species and decay class
}
\description{
Decay Class Reduction Factor (DCRF)
}
\examples{
NA
}
