% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HARMON_2011.R
\docType{data}
\name{HARMON_2011}
\alias{HARMON_2011}
\title{Data: Relative decay reduction factors}
\format{
A data frame with 261 rows and 17 columns:
\describe{
\item{Genus}
\item{Species}
\item{Code}
\item{NFI_CODE}{Adapted species code to match NFI naming}
...
}
}
\source{
Adapted from Harmon et al., 2011 Appendix D.
Added Values:
\itemize{
\item Populus trichoptera added. Same values as Populus balsamifera
}
}
\usage{
HARMON_2011
}
\description{
Data: Relative decay reduction factors
}
\keyword{datasets}
