% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JENKINS_GROUPS.R
\docType{data}
\name{JENKINS_GROUPS}
\alias{JENKINS_GROUPS}
\title{Data: NFI species code to Jenkins species group mapping}
\format{
A data frame with 43 rows and 2 columns:
\describe{
\item{SPECIES}{NFI species code}
\item{SPECIES_GROUP}{Corresponding Jenkins et al. (2003) species group}
}
}
\source{
Species group assignments based on Jenkins et al. (2003) National-scale
biomass estimators for United States tree species. Forest Science, 49(1): 12-35.
}
\usage{
JENKINS_GROUPS
}
\description{
Data: NFI species code to Jenkins species group mapping
}
\keyword{datasets}
