% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkinsCalculator.R
\name{jenkinsCalculator}
\alias{jenkinsCalculator}
\title{Internal Function: Jenkins Calculator}
\usage{
jenkinsCalculator(data, dbh, species, appearance = NULL, output, decay)
}
\arguments{
\item{data}{User specified dataframe.}

\item{dbh}{Column within data where DBH is specified.}

\item{species}{Column within data where species is specified. NFI codes.}

\item{appearance}{Column within data where tree appearance is specified.}

\item{output}{Either "biomass" or "carbon".}

\item{decay}{Logical. Should the decay class reduction factor be applied?}
}
\value{
A vector
}
\description{
Internal Function: Jenkins Calculator
}
\keyword{internal}
