% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungEqn.R
\name{ungEqn}
\alias{ungEqn}
\title{Internal Function: Allometric biomass equation (Ung et al., 2008)}
\usage{
ungEqn(data, dbh, height = NULL, beta_list, i)
}
\arguments{
\item{data}{User specified dataframe.}

\item{dbh}{Column within data where DBH is specified.}

\item{height}{Column within data where Height is specified. NULL for DBH-only equations.}

\item{beta_list}{Beta values specified from parent function.}

\item{i}{Row index for iterating.}
}
\value{
Numeric value.
}
\description{
Internal Function: Allometric biomass equation (Ung et al., 2008)
}
\examples{
NA
}
