% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSpecies.R
\name{validateSpecies}
\alias{validateSpecies}
\title{Internal Function: Validate Species Input}
\usage{
validateSpecies(species_input, eval)
}
\arguments{
\item{species_input}{Vector of species codes from user data.}

\item{eval}{The allometric equation method string.}
}
\value{
NULL invisibly; aborts if species are invalid.
}
\description{
Internal Function: Validate Species Input
}
