% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodCalculator.R
\name{woodCalculator}
\alias{woodCalculator}
\title{Internal Function: Wood Calculator helper function}
\usage{
woodCalculator(
  data,
  method,
  output,
  dbh,
  height = NULL,
  species,
  func,
  appearance = NULL,
  decay = TRUE
)
}
\arguments{
\item{data}{User specified dataframe.}

\item{method}{Dataframe within the package to retrive beta values from.
Options are "LAMBERT_1", "LAMBERT_2", "UNG_1", "UNG_2".}

\item{output}{Either "biomass" or "carbon". Biomass gives values in Kg. Carbon gives values in Mg/ha.}

\item{dbh}{Column within data where Diameter at Breast Height (dbh) is specified.}

\item{height}{Column within data where Height is specified.}

\item{species}{Column within data where species is specified. NFI codes for species.}

\item{func}{Calls function ungEqn.}

\item{appearance}{Column within data where tree appearance is specified.}

\item{decay}{Logical with default = TRUE. Should the decay class reduction factor be applied?}
}
\value{
A vector
}
\description{
Internal Function: Wood Calculator helper function
}
\examples{
NA
}
\keyword{internal}
